+++ packages/sdk/react-shell/src/components/InvitationList/InvitationList.stories.tsx
@@ -10,7 +10,7 @@ import React, { useMemo, useState } from 'react';
 import { ThemeContext, useThemeContext } from '@dxos/aurora';
 import { mx, osTx } from '@dxos/aurora-theme';
 import { CancellableInvitationObservable, Client, Invitation } from '@dxos/client';
-import { TestBuilder } from '@dxos/client-services/testing';
+import { TestBuilder } from '@dxos/client/testing';
 import { log } from '@dxos/log';
 import { Loading } from '@dxos/react-appkit';
 import { useAsyncEffect } from '@dxos/react-async';